Fix validation of overlays
Correct the case-clause which validates the overlay tuple. The method
was designed for tuples with the format `{:copy, ["src", "target"]}`
but `Mix.Releases.Overlays` only accepts tuples which don't carry the
options in a list. This resulted in a ArgumentError, which caused an
`Mix.Releases.Config.LoadError`, which threw an error message
in lib/distillery/tasks/release.ex at line 73.

Now the case clause matches overlay tuples with one or two arguments
and therefore follows the spec in `Mix.Releases.Overlays`.