Fix plugins only specified in an environment
When specifying a plugin in an environment only (and not a release) the
plugin is prepended to the profile.plugins attribute. However since
this defaults to nil, the result is an improper list in the format:

    [SampleApp.TestPlugin | nil]

This commit fixes it so that the profile.plugins defaults to an empty
list, resulting in:

    [SampleApp.TestPlugin | []]

A test has been added which loads a sample plugin from the rel
directory and verifies that the list contains a list with the plugin.
This test will fail when the list is an improper list.

When merging the profile and environment, there is a check in place so
that lists are correctly merged together.

The integration test has been modified to ensure that all the plugins
run.