Ensure that VMARGS_PATH/SYS_CONFIG_PATH handling is re-entrant
If users were setting VMARGS_PATH/SYS_CONFIG_PATH directly, the previous
handling of these env vars would result in incorrect behavior when used
with the `start` task, or in any instance where the boot script was
re-entrant (such as calling the boot script from a hook or custom
command). This fix ensures that this handling behaves correctly in these
scenarios.