Set name of process to the script path
This addresses the situation where looking for the running release in
the output of `ps ax` shows the path to `beam.smp` with a bunch of
arguments, rather than the path of the release script which is being
executed. If you have multiple releases running, the former is very
difficult to read and find the pid of a given release process, with this
change it is now just a matter of finding the process with the path you
care about.