Add support for Mix.Config and custom config providers
This commit provides initial support for config.exs, as well as custom
configuration providers beyond what Mix.Config provides. Currently this
manifests as implementing a behaviour defining two callbacks, `init/1` and
`get/1`. The latter is currently not used, but I have future plans which
will likely need it, so implementations should treat that callback as a
function for fetching the configured value for a specific key. See the
docs for more info.

An example of this would look like so:

    defmodule SystemEnvProvider do
      use Mix.Releases.Config.Provider

      def init(_), do: :ok
      def get(keypath) do
        keypath
        |> Enum.map(&Atom.to_string/1)
        |> Enum.map(&String.upcase/1)
        |> Enum.join("_")
        |> System.get_env
      end
    end

You would then configure your release to use this provider like so:

    environment :prod do
      set config_providers: [SystemEnvProvider]
    end

The default provider for config.exs files will always come first, and
any providers you specify here will be applied after config.exs is
evaluated.

If you look at how the Elixir provider is built, the bulk of what it
does happens in `init/1`, this is because `init/1` is called on startup,
and the provider is configured by default to receive a path to the
generated config.exs file containing the merged config files from your
project. Subsequent calls to `get/1` just need to check the application
environment.

This design is subject to change, but hopefully will be shown to work in
practice.