Raise error when ERTS was compiled for another OS
This is a fairly naive check, but tests the detected erl and if it
fails to run a simple command, is considered unusable and thus an error
is raised and shown to the user.

Perhaps in the future we can get more clever, but this is about as
precise as I want to be with it at the moment.

Closes #370