Do not specify start type for :mix
This conflicts with some libraries which incorrectly add :mix to their
extra_applications list. Distillery only wants Mix code to be available,
but doesn't actually need it started, but if a dependency of the
releasing project specifies :mix in it's applications, it will result in
the release only partially booting, without any errors. There is now a
warning for this situation, but this commit addresses the fact that
Distillery hardcodes the start type (though one could override it in
their `rel/config.exs`, I've deemed it unnecessary to specify the start
type, for now).

Closes #443 as well.