Add link to distillery-aws-example repo reference
When I was skimming through the 'deploying to aws' guide it took me longer than expected to find the link to the distillery-aws-example repo after I saw it referenced in the 'Provisioning our pipeline' section. Thinking we could just make the reference a link to save other people a little bit of time?