Fix grep command
According to grep manual of both OpenBSD and macOS:

grep is used for simple patterns and basic regular expressions (BREs);
egrep can handle extended regular expressions (EREs).

"|" is an ordinary character of BREs, so `^\-\(sn\|n\)ame` should not work
for grep. However, it works in macOS, which lead to the issue I
guess. Gonna need to be careful to the commands in macOS.