Do not load configs for secondary commands
Secondary commands are those which are not used to start a release or
manipulate a release in such a way that configuration changes are
required to be made. These commands should use
configuration that has already been generated by the running release.

Examples:

- `start`, `console`, `foreground` - primary commands, they are
responsible for generating the configs they use
- `ping`, a secondary command, needs to contact a running node, and so needs the node name
and cookie, which can be read from `vm.args` without needing to generate
a fresh copy from the source config, which may be destructive and can
conflict with other commands
- `reload_config`, a primary command because it needs to reload configs
from source and generate new on-disk copies

This is a potentially breaking command, but the fix is easy while the
bugs the current behavior causes is much nastier and more painful.

Fixes #604