Quote the boot script with single quotes (#676)
Using unescaped double quotes within double quotes didn't seem to be working as expected, as I was having issues running a batch file produced from this template in a directory which contained spaces. This is similar to how it was suggested that powershell be called from a batch file here: https://stackoverflow.com/questions/19335004/how-to-run-a-powershell-script-from-a-batch-file.